+++ govtool/frontend/src/pages/DRepDirectoryContent.tsx
@@ -46,7 +46,7 @@ export const DRepDirectoryContent: FC<DRepDirectoryContentProps> = ({
   const { dRepID: myDRepId, pendingTransaction, stakeKey } = useCardano();
   const { t } = useTranslation();
   const { debouncedSearchText, ...dataActionsBarProps } = useDataActionsBar();
-  const { chosenFilters, chosenSorting, setChosenSorting } =
+  const { chosenFilters, chosenSorting, setChosenFilters, setChosenSorting } =
     dataActionsBarProps;
 
   const [inProgressDelegationDRepData, setInProgressDelegationDRepData] =
@@ -54,7 +54,8 @@ export const DRepDirectoryContent: FC<DRepDirectoryContentProps> = ({
 
   useEffect(() => {
     if (!chosenSorting) setChosenSorting(DRepListSort.Random);
-  }, [chosenSorting, setChosenSorting]);
+    if (!chosenFilters.length) setChosenFilters([DRepStatus.Active]);
+  }, [chosenSorting, setChosenSorting, chosenFilters, setChosenFilters]);
 
   const { delegate, isDelegating } = useDelegateTodRep();